Move ConveryKey helper to utils
